给自己合并 play 主分支

+++ framework/pym/play/commands/intellij.py
@@ -46,18 +46,20 @@ def execute(**kargs):
             srcpath = os.path.join(module, 'src')
             lXML += '        <content url="file://%s">\n            <sourceFolder url="file://%s" isTestSource="false" />\n        </content>\n' % (module, os.path.join(module, 'app').replace('\\', '/'))
             if os.path.exists(srcpath):
-                msXML += '                    <root url="file://$MODULE_DIR$%s"/>\n' % (app.toRelative(srcpath).replace('\\', '/'))
+                msXML += '                    <root url="file://$MODULE_DIR$/%s"/>\n' % (app.toRelative(srcpath).replace('\\', '/'))
             if os.path.exists(libpath):
-                mlXML += '                    <root url="file://$MODULE_DIR$%s"/>\n' % (app.toRelative(libpath).replace('\\', '/'))
-                jdXML += '                <jarDirectory url="file://$MODULE_DIR$%s" recursive="false"/>\n' % (app.toRelative(libpath).replace('\\', '/'))
+                mlXML += '                    <root url="file://$MODULE_DIR$/%s"/>\n' % (app.toRelative(libpath).replace('\\', '/'))
+                jdXML += '                <jarDirectory url="file://$MODULE_DIR$/%s" recursive="false"/>\n' % (app.toRelative(libpath).replace('\\', '/'))
     replaceAll(imlFile, r'%LINKS%', lXML)
     replaceAll(imlFile, r'%MODULE_LINKS%', mlXML)
     replaceAll(imlFile, r'%MODULE_LIB_CLASSES%', msXML)
     replaceAll(imlFile, r'%MODULE_LIBRARIES%', jdXML)
     
     iprFile = os.path.join(app.path, application_name + '.ipr')
-    shutil.copyfile(os.path.join(play_env["basedir"], 'resources/idea/iprTemplate.xml'), iprFile)
-    replaceAll(iprFile, r'%PROJECT_NAME%', application_name)
+    # Only copy/create if missing to avoid overwriting customizations
+    if not os.path.exists(iprFile):
+        shutil.copyfile(os.path.join(play_env["basedir"], 'resources/idea/iprTemplate.xml'), iprFile)
+        replaceAll(iprFile, r'%PROJECT_NAME%', application_name)
     
 
     print "~ OK, the application is ready for Intellij Idea"
